+++ main.rb
@@ -7,6 +7,10 @@ require "discorb"
 require "dotenv"
 require "etc"
 require "log4r"
+require "mongo"
+require_relative "config/perms"
+database = Mongo::Client.new("mongodb://127.0.0.1:27017/distools")
+puts "(Maybe) connected to Mongo."
 log = Log4r::Logger.new("discorb")
 log.add Log4r::FileOutputter.new("filelog",
                                  filename: "distools.log",
@@ -16,21 +20,17 @@ log.add Log4r::IOOutputter.new("stderr", $stderr,
                                level: Logger::WARN)
 systemUsername = Etc.getpwuid(Process.uid).name
 uname = Etc.uname
-# Arrays (blacklist, owners, friend guilds, etc)
-botOwners = [393971637642461185,734740501798060092]
-superBotOwners = [393971637642461185]
-blacklist = []
 friendGuilds = [961703572741951518,972546971460059197]
 botDescription = "Hosted instance of distools used for development."
-botVersion = "1.3"
+botVersion = "2.1 RW"
 commit = `git rev-parse --short HEAD`
 commitMsg = `git show-branch --no-name`
 blacklistContactPoint = "<@393971637642461185>"
 
 File.foreach("index.txt") { |line| puts line }
-puts "loaded bot owners: #{botOwners}"
-puts "loaded blacklist: #{blacklist}"
-puts "loaded friend servers: #{friendGuilds}"
+puts "loaded bot owners: #{BotPerms::BOTOWNERS}"
+puts "loaded blacklist: #{BotPerms::BLACKLISTED}"
+puts "loaded friend guilds: #{friendGuilds}"
 Dotenv.load  # Loads .env file
 puts "loaded .env and thus i am proceeding"
 client = Discorb::Client.new(logger: log)  # Create client for connecting to Discord
@@ -57,7 +57,7 @@ end
 # guild whois
 client.slash("guild", "returns guild information") do |interaction|
 # Before absolutely anything run the Blacklist Test
- if blacklist.include?(interaction.target.id) === true
+ if BotPerms::BLACKLISTED.include?(interaction.target.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{blacklistContactPoint} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
   if friendGuilds.include?(interaction.guild.id) === true
@@ -79,7 +79,7 @@ end
 end
 # invite command
 client.slash("invite", "gives you 2 invite links: limited and full") do |interaction|
- if blacklist.include?(interaction.target.id) === true
+ if BotPerms::BLACKLISTED.include?(interaction.target.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{blacklistContactPoint} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
   interaction.post(embed: Discorb::Embed.new("Invite me!", "I have two invite links. One is the limited version and one is full admin permissions.\n\n[Limited](https://discord.com/api/oauth2/authorize?client_id=#{ENV['ID']}&permissions=1634771397751&scope=applications.commands%20bot)\n[Full](https://discord.com/api/oauth2/authorize?client_id=#{ENV['ID']}&permissions=8&scope=applications.commands%20bot)", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
@@ -87,15 +87,15 @@ end
 end
 
 client.slash("instance", "Details about this instance of distools including owners, blacklist, and more") do |interaction|
- if blacklist.include?(interaction.target.id) === true
+ if BotPerms::BLACKLISTED.include?(interaction.target.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{blacklistContactPoint} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
-  interaction.post(embed: Discorb::Embed.new("About this instance", "The `botOwners` array: #{botOwners}\nThe `blacklist` array: #{blacklist}\nThe `friendGuilds` array: #{friendGuilds}\n\nThe description of this instance is: #{botDescription}\ndistools version: #{botVersion}\n#{client.user} is on distools commit #{commit} with the commit message:\n```\n#{commitMsg}```", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+  interaction.post(embed: Discorb::Embed.new("About this instance", "The `botOwners` array: #{BotPerms::botOwners}\nThe `blacklist` array: #{BotPerms::blacklist}\nThe `friendGuilds` array: #{friendGuilds}\n\nThe description of this instance is: #{botDescription}\ndistools version: #{botVersion}\n#{client.user} is on distools commit #{commit} with the commit message:\n```\n#{commitMsg}```", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
 end
 end
 
 client.slash("hello", "bot stats") do |interaction|
- if blacklist.include?(interaction.target.id) === true
+ if BotPerms::BLACKLISTED.include?(interaction.target.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{blacklistContactPoint} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
   interaction.post(embed: Discorb::Embed.new("distools #{botVersion} | hello world :)", "I am Distools, a bot made by Winfinity\#1252.", color: Discorb::Color.from_rgb(201, 0, 0), fields: [Discorb::Embed::Field.new("ruby ver", "`#{RUBY_VERSION}-p#{RUBY_PATCHLEVEL}`", inline: false), Discorb::Embed::Field.new("discorb ver", "`#{Discorb::VERSION}`", inline: false), Discorb::Embed::Field.new("OS", "`#{uname[:sysname]} #{uname[:release]}`", inline: false), Discorb::Embed::Field.new("running as user", "`#{systemUsername}`", inline: false)]), ephemeral: false)
@@ -103,7 +103,7 @@ end
 end
 
 client.slash("help", "gives you my commands") do |interaction|
- if blacklist.include?(interaction.target.id) === true
+ if BotPerms::BLACKLISTED.include?(interaction.target.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{blacklistContactPoint} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
   interaction.post(embed: Discorb::Embed.new("Help | distools #{botVersion}", "I am Distools. A bot made with :heart: (+ caffiene and Ruby) by Winfinity#1252.\n\n**Command List**\n`/hello` - (Interaction) Gives you a list of bot stats.\n`/instance` - (Interaction) Gives you information about this distools instance.\n`/help` - (Interaction) Lists commands.\n`/stop` - (Interaction, bot owners only) Stops the bot using `exit`.\n`/guild` - (Interaction) Guild information.\n`/whoami` - (Interaction) You are #{interaction.target.name}. Seriously, this gives you more information.\n`/why-god-why` - (Interaction, bot owners only) Tests everything. Template command.\n`dist.eval` - (Bot owners only) Evaluates a piece of ruby code.\n`dist.exec` - (\"Super\" bot owners only) Executes a script on the machine distools is running on (via the backticks)\n`/am-i-a-dev` - (Interaction) Tells you whether you are listed as a dev or not.\n`/invite` - (Interaction) Returns invite links.\n\nDistools is open source (the code is [right here](https://github.com/haydenwalker980)) - meaning you are free to use, modify, and distribute it. If you enjoy distools and want to keep the lights on when we inevitably move to a server, [please consider contributing](https://liberapay.com/Win) via liberapay. To have your donation count against an indicator in user information, leave your Discord ID in your liberapay description.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
@@ -117,7 +117,7 @@ client.on :message do |message|
   puts "Message in #{message.guild.name} (#{message.guild.id}) from #{message.author.name} (#{message.author.id}): #{message.content}"
   next if message.author.bot?
   next unless message.content.start_with?("dist.eval ")
-  unless botOwners.include?(message.author.id)
+  unless BotPerms::BOTOWNERS.include?(message.author.id)
     message.reply embed: Discorb::Embed.new("Uhhh", "No.", color: Discorb::Color.from_rgb(201, 0, 0))
     next
   end
@@ -138,7 +138,7 @@ end
 
 client.on :message do |message|
   next unless message.content.start_with?("dist.exec ")
-  unless superBotOwners.include?(message.author.id)
+  unless BotPerms::EXECUSERS.include?(message.author.id)
         message.reply("Yeah mate since this is such a high risk operation only the super bot owner can run it.")
         next
   end
@@ -186,7 +186,7 @@ end
 # Gonna see if this works. Ctrl C does not seem like a good way to end a bot so I
 # think I will gracefully stop it
 client.slash("stop", "[owner only] shutting down") do |interaction|
-  if botOwners.include?(interaction.target.id) === true
+  if BotPerms::BOTOWNERS.include?(interaction.target.id) === true
 	puts "peaceful shutdown initiated"
 	interaction.post("ok, shutting down, goodbye :sleeping:", ephemeral: false)
 	exit
@@ -196,11 +196,11 @@ client.slash("stop", "[owner only] shutting down") do |interaction|
 end
 
 client.slash("am-i-a-dev", "tells you whether the bot thinks you are a dev or not") do |interaction|
- if blacklist.include?(interaction.target.id) === true
+ if BotPerms::BLACKLISTED.include?(interaction.target.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{blacklistContactPoint} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
- if botOwners.include?(interaction.target.id) === true
-	if superBotOwners.include?(interaction.target.id) === true
+ if BotPerms::BOTOWNERS.include?(interaction.target.id) === true
+	if BotPerms::EXECUSERS.include?(interaction.target.id) === true
 		interaction.post(embed: Discorb::Embed.new("The results are in!!", "You are an owner. Additionally, you have permission to run `dist.exec`.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
 	else
 		interaction.post(embed: Discorb::Embed.new("The results are in!!", "You are an owner, so you have permissions to run some high level command, but you are unable to run `dist.exec`.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
@@ -212,7 +212,7 @@ end
 end
 
 client.slash("why-god-why", "[owner only] tests ABSOLUTELY EVERYTHING. could serve as a template command") do |interaction|
-  if botOwners.include?(interaction.target.id) === true
+  if BotPerms::BOTOWNERS.include?(interaction.target.id) === true
 	puts "Oh dear god it's a stress test. We're all gonna die."
 	interaction.post("*insert :hollow: here*", embed: Discorb::Embed.new("embed test :hollow:", "This is only a test"), ephemeral: false)
   else